sched: fair: Utilize more higher capacity cpus in asymmetric systems
The current fbt starts the CPU search from the suitable
sched_group/cluster. If an eligible CPU (either busy or idle) is found
in this primary cluster, we break and use this eligible CPU. The search
is expanded to the other cluster only when either an eligible CPU is
not found or boost is enabled. This works for the existing topologies
like 6 + 2 or 4 + 3 + 1 systems. However with a topology like 6 + 1 + 1
where the mid capacity cluster has only 1 CPU, this policy results in
excessive packing on that single CPU. So expand the search to the higher
capacity CPU as well when the start cluster is mid capacity cluster.
The exception is when the mid capacity CPU is idle. This is a power
optimization to not to use the higher capacity CPU more.

Change-Id: I15c591dc2f61fbef203545c86e038a0ff918806c
Signed-off-by: Lingutla Chandrasekhar <clingutla@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>