op_touchscreen: Avoid -unknown-warning-option(s) when compiling with Clang
Subsequent to 4cc2001 ("op_touchscreen: major clean-up")

Clang does not recognize unused-but-set-variable or multilevel
unused-const-variable as valid flags. The latter only exists as a
singular unused-const-variable option which is implicitly enabled
alongside -unused-variable. Thus, these should only be passed if we
are compiling with GCC (cc != clang) in order to avoid potential
build interruptions due to unrecognized CFLAGs.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>