neutrino: Enable LTO_CLANG and dependencies
I'm choosing to squash all LTO related configs into one commit for
the sake of others who may be attempting to compile this kernel and
experience toolchain related complications as certain LLVM features
can prove to be a hassle for some who's build environments are not
properly equipped to handle them. If you happen to be one of these
people, simply revert this commit and carry on.

Otherwise this acts as a switch to enable support for Clang LTO as
well as linking the with LLD (f399931), which is a prerequisite for
enabling RELR Relocations (ad30f81).

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>