dsi-panel-samsung_oneplus_dsc: disable dsc compression for 1080p
As stated by kdrag0n: VESA claims this is low-latency and *visually*
lossless, but not *fully* lossless. Why take the extra latency and
quality loss, as small as it is, if we have the bandwidth to
transfer the data without compression?

Note: for 1080p only due to 1440@90 needing dsc compression