gulch_defconfig: switch to cubic by default
Let users choose experimental bbr2 if they like.

I found some old recreatable issues on aospa's kernel which could be
recreated on older bbr here, but not as wildly on bbr2. With that
said, there is still some inconsistent behavior depending on network
type with bbr2 that doesn't occur on good old cubic.